(WIP) Css 1272/create machine operator (#814)
* Fixes openapi 3.0 refs.

The refs were wrong as most things were referred to as
'params', which they weren't. Updated most to be body and fixed one
broken endpoint "GetSyncReports" as the path param
was missing. This needs testing but I presume it should be ok.

* Change reports since param into query

This param is to be optional, and should not be a path param due to this.

* Initial commit

* Start on charm db relations

* Action and schema upgrade additions

* WIP

- Updates charm to use pgsql ops lib
- Handles schema upgrades
- Handles statuses
- Handles snap state

* Handle statuses and validate schema upgrade

* Handle add/append users action

- Uses binary pyth package for bcrypt.
- Updates config path for snap to be static, no need
to dynamically replace it. It'll never change.
- Adds the action.

* Add restart action

* Refactor actions to use TYPE_CHECKING

This reduces class bloat, making it easier
to digest exactly what your actions, lifecycle hooks and relations are.

Will move the rest into neatly separated py modules too soon.

* Status updates for clear
actions

Additionally updates all config to have descriptions and defaults.

Removes unncessary config that on-prem users will not use.

* Charm complete

* Initial integration setup

* Initial integration test utils

* Fixture deployment bundle

Fixtures the deployment bundle in a variety of ways.
Allowing us to test across different scopes of integration environments.
Each is capable of using a different rendered bundle too.

* More small bits of progress...

* More progress...

* Note for self.

* Remove need to build snaps over and over

* Fix comments